Minor bug fixes

+++ src/main/java/com/github/pavelvashkevich/bankmicroservice/repository/BankAccountRepository.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 
 public interface BankAccountRepository extends JpaRepository<BankAccount, Long> {
 
-    Optional<BankAccount> findByAccountNumber(int accountNumber);
+    Optional<BankAccount> findByAccountNumber(long accountNumber);
 }
