+++ src/test/java/com/github/pavelvashkevich/bankmicroservice/repository/ClientRepositoryTest.java
@@ -16,8 +16,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
 public class ClientRepositoryTest {
-    private static final int TEST_BANK_ACCOUNT_1_NUM = 1234567891;
-    private static final int TEST_BANK_ACCOUNT_2_NUM = 1234567892;
+    private static final long TEST_BANK_ACCOUNT_1_NUM = 1234567891L;
+    private static final long TEST_BANK_ACCOUNT_2_NUM = 1234567892L;
     private static final LocalDate TEST_DATE = LocalDate.of(2000, 1, 1);
 
     private static Client client;
